+++ GFramework.Cqrs.Tests/Cqrs/CqrsDispatcherCacheTests.cs
@@ -224,124 +224,8 @@ internal sealed class CqrsDispatcherCacheTests
     /// </summary>
     private static async Task DrainAsync<T>(IAsyncEnumerable<T> stream)
     {
-        await foreach (var _ in stream)
+        await foreach (var _ in stream.ConfigureAwait(false))
         {
         }
-    }
-}
-
-/// <summary>
-///     用于验证 request 服务类型缓存的测试请求。
-/// </summary>
-internal sealed record DispatcherCacheRequest : IRequest<int>;
-
-/// <summary>
-///     用于验证 notification 服务类型缓存的测试通知。
-/// </summary>
-internal sealed record DispatcherCacheNotification : INotification;
-
-/// <summary>
-///     用于验证 stream 服务类型缓存的测试请求。
-/// </summary>
-internal sealed record DispatcherCacheStreamRequest : IStreamRequest<int>;
-
-/// <summary>
-///     用于验证 pipeline invoker 缓存的测试请求。
-/// </summary>
-internal sealed record DispatcherPipelineCacheRequest : IRequest<int>;
-
-/// <summary>
-///     用于验证按响应类型分层 request invoker 缓存的测试请求。
-/// </summary>
-internal sealed record DispatcherStringCacheRequest : IRequest<string>;
-
-/// <summary>
-///     处理 <see cref="DispatcherCacheRequest" />。
-/// </summary>
-internal sealed class DispatcherCacheRequestHandler : IRequestHandler<DispatcherCacheRequest, int>
-{
-    /// <summary>
-    ///     返回固定结果，供缓存测试验证 dispatcher 请求路径。
-    /// </summary>
-    public ValueTask<int> Handle(DispatcherCacheRequest request, CancellationToken cancellationToken)
-    {
-        return ValueTask.FromResult(1);
-    }
-}
-
-/// <summary>
-///     处理 <see cref="DispatcherCacheNotification" />。
-/// </summary>
-internal sealed class DispatcherCacheNotificationHandler : INotificationHandler<DispatcherCacheNotification>
-{
-    /// <summary>
-    ///     消费通知，不执行额外副作用。
-    /// </summary>
-    public ValueTask Handle(DispatcherCacheNotification notification, CancellationToken cancellationToken)
-    {
-        return ValueTask.CompletedTask;
-    }
-}
-
-/// <summary>
-///     处理 <see cref="DispatcherCacheStreamRequest" />。
-/// </summary>
-internal sealed class DispatcherCacheStreamHandler : IStreamRequestHandler<DispatcherCacheStreamRequest, int>
-{
-    /// <summary>
-    ///     返回一个最小流，供缓存测试命中 stream 分发路径。
-    /// </summary>
-    public async IAsyncEnumerable<int> Handle(
-        DispatcherCacheStreamRequest request,
-        [EnumeratorCancellation] CancellationToken cancellationToken)
-    {
-        yield return 1;
-        await Task.CompletedTask;
-    }
-}
-
-/// <summary>
-///     处理 <see cref="DispatcherPipelineCacheRequest" />。
-/// </summary>
-internal sealed class DispatcherPipelineCacheRequestHandler : IRequestHandler<DispatcherPipelineCacheRequest, int>
-{
-    /// <summary>
-    ///     返回固定结果，供 pipeline 缓存测试使用。
-    /// </summary>
-    public ValueTask<int> Handle(DispatcherPipelineCacheRequest request, CancellationToken cancellationToken)
-    {
-        return ValueTask.FromResult(2);
-    }
-}
-
-/// <summary>
-///     处理 <see cref="DispatcherStringCacheRequest" />。
-/// </summary>
-internal sealed class DispatcherStringCacheRequestHandler : IRequestHandler<DispatcherStringCacheRequest, string>
-{
-    /// <summary>
-    ///     返回固定字符串，供按响应类型缓存测试验证 string 路径。
-    /// </summary>
-    public ValueTask<string> Handle(DispatcherStringCacheRequest request, CancellationToken cancellationToken)
-    {
-        return ValueTask.FromResult("dispatcher-cache");
-    }
-}
-
-/// <summary>
-///     为 <see cref="DispatcherPipelineCacheRequest" /> 提供最小 pipeline 行为，
-///     用于命中 dispatcher 的 pipeline invoker 缓存分支。
-/// </summary>
-internal sealed class DispatcherPipelineCacheBehavior : IPipelineBehavior<DispatcherPipelineCacheRequest, int>
-{
-    /// <summary>
-    ///     直接转发到下一个处理器。
-    /// </summary>
-    public ValueTask<int> Handle(
-        DispatcherPipelineCacheRequest request,
-        MessageHandlerDelegate<DispatcherPipelineCacheRequest, int> next,
-        CancellationToken cancellationToken)
-    {
-        return next(request, cancellationToken);
     }
 }
