+++ GFramework.Cqrs.Tests/Cqrs/DispatcherCacheRequestHandler.cs
@@ -1,0 +1,21 @@
+using System.Threading;
+using GFramework.Cqrs.Abstractions.Cqrs;
+
+namespace GFramework.Cqrs.Tests.Cqrs;
+
+/// <summary>
+///     处理 <see cref="DispatcherCacheRequest" />。
+/// </summary>
+internal sealed class DispatcherCacheRequestHandler : IRequestHandler<DispatcherCacheRequest, int>
+{
+    /// <summary>
+    ///     返回固定结果，供缓存测试验证 dispatcher 请求路径。
+    /// </summary>
+    /// <param name="request">当前请求。</param>
+    /// <param name="cancellationToken">取消令牌。</param>
+    /// <returns>固定整数结果。</returns>
+    public ValueTask<int> Handle(DispatcherCacheRequest request, CancellationToken cancellationToken)
+    {
+        return ValueTask.FromResult(1);
+    }
+}
